refactor : use-chat-room파일명 변경  타입 단언->제네릭 사용 #10
