+++ src/components/chat/message-item.tsx
@@ -11,7 +11,7 @@ interface Props {
 }
 
 export function MessageItem({ message, isOwn }: Props) {
-  const { data: members = [] } = useRoomMembers(message.chat_room_id)
+  const { data: members = [] } = useRoomMembers(message.roomId)
 
   const member = members.find((m) => m.user_id === message.user_id)
   const displayName = member?.user?.nickname?.trim() || message.user_id.slice(0, 8)
