+++ src/hooks/use-room-members.ts
@@ -6,10 +6,10 @@ import type { RoomMemberQuery } from "@/types/chatroommember";
 export function useRoomMembers(roomId: string) {
   const supabase = createClient();
 
-  return useQuery({
+  return useQuery<RoomMemberQuery[]>({
     queryKey: ["room-members", roomId],
     enabled: !!roomId,
-    queryFn: async (): Promise<RoomMemberQuery[]> => {
+    queryFn: async () => {
       const { data, error } = await supabase
         .from("chatroommember")
         .select("chat_room_id, user_id, created_at, user:user_id(nickname)")
@@ -18,7 +18,7 @@ export function useRoomMembers(roomId: string) {
 
       if (error) throw error;
 
-      return data as RoomMemberQuery[];
+      return data as unknown as RoomMemberQuery[];
     },
   });
 }
