--- src/hooks/use-room.ts
@@ -1,24 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-
-import { createClient } from "@/lib/supabase/client";
-import type { Room } from "@/types/chatroom";
-
-export function useRoom(roomId: string) {
-  const supabase = createClient();
-
-  return useQuery({
-    queryKey: ["room", roomId],
-    enabled: !!roomId,
-    queryFn: async (): Promise<Room> => {
-      const { data, error } = await supabase
-        .from("chatroom")
-        .select("*")
-        .eq("id", roomId)
-        .single();
-
-      if (error) throw error;
-
-      return data as Room;
-    },
-  });
-}
